refractor(backend): minimize repeated code
