+++ chatbot-core/backend/app/main.py
@@ -9,9 +9,8 @@ from app.routers import base
 from app.routers.v1 import agent
 from app.routers.v1 import chat
 from app.routers.v1 import connector
-from app.routers.v1 import embedding
 from app.routers.v1 import folder
-from app.routers.v1 import llm
+from app.routers.v1 import provider
 from app.routers.v1 import user
 from app.settings import Constants
 from app.utils.api.helpers import get_logger
@@ -74,8 +73,7 @@ def create_app() -> FastAPI:
     app.include_router(router=folder.router, prefix=Constants.FASTAPI_PREFIX)
     app.include_router(router=agent.router, prefix=Constants.FASTAPI_PREFIX)
     app.include_router(router=user.router, prefix=Constants.FASTAPI_PREFIX)
-    app.include_router(router=llm.router, prefix=Constants.FASTAPI_PREFIX)
-    app.include_router(router=embedding.router, prefix=Constants.FASTAPI_PREFIX)
+    app.include_router(router=provider.router, prefix=Constants.FASTAPI_PREFIX)
 
     return app
