+++ chatbot-core/backend/app/models/__init__.py
@@ -2,10 +2,10 @@ from app.models.agent import Agent
 from app.models.chat import ChatFeedback
 from app.models.chat import ChatMessage
 from app.models.chat import ChatSession
-from app.models.embedding import EmbeddingProvider
 from app.models.folder import Folder
-from app.models.llm import LLMProvider
 from app.models.prompt import Prompt
+from app.models.provider import EmbeddingProvider
+from app.models.provider import LLMProvider
 from app.models.user import User
 from app.models.user import UserSetting
